split out embed iframe heigh logic from atom iframe
