+++ src/web/browser/embedIframe/embedIframe.ts
@@ -1,0 +1,4 @@
+import { updateIframeHeight } from '../updateIframeHeight';
+
+export const embedIframe = (): Promise<void> =>
+    updateIframeHeight('.embed__iframe');
