+++ src/web/browser/embedIframe/init.ts
@@ -1,0 +1,5 @@
+import '../webpackPublicPath';
+import { startup } from '@root/src/web/browser/startup';
+import { embedIframe } from './embedIframe';
+
+startup('embedIframe', null, embedIframe);
