+++ src/web/components/elements/UnsafeEmbedBlockComponent.tsx
@@ -36,7 +36,7 @@ const unsafeEmbedWrapperStyles = css`
 export const UnsafeEmbedBlockComponent = ({ html, alt, index }: Props) => (
     <figure className={unsafeEmbedWrapperStyles}>
         <iframe
-            className={`${fullWidthStyles} atom__iframe`}
+            className={`${fullWidthStyles} embed__iframe`}
             title={alt}
             // name is used to identify each unique iframe on the page to resize
             // we therefore use the "unsafe-embed-" prefix followed by index to
