+++ src/web/layouts/Standard.stories.tsx
@@ -17,7 +17,7 @@ import { MatchReport } from '@root/fixtures/articles/MatchReport';
 import { NAV } from '@root/fixtures/NAV';
 
 import { HydrateApp } from '@root/src/web/components/HydrateApp';
-import { atomIframe } from '@root/src/web/browser/atomIframe/atomIframe';
+import { embedIframe } from '@root/src/web/browser/embedIframe/embedIframe';
 import { mockRESTCalls } from '@root/src/web/lib/mockRESTCalls';
 
 import { DecideLayout } from './DecideLayout';
@@ -49,7 +49,7 @@ const HydratedLayout = ({ ServerCAPI }: { ServerCAPI: CAPIType }) => {
     useEffect(() => {
         const CAPI = makeGuardianBrowserCAPI(ServerCAPI);
         HydrateApp({ CAPI, NAV });
-        atomIframe();
+        embedIframe();
     }, [ServerCAPI]);
     return <DecideLayout CAPI={ServerCAPI} NAV={NAV} />;
 };
